+++ apps/veil/CHANGELOG.md
@@ -1,5 +1,28 @@
 # penumbra-veil
 
+## 0.1.1
+
+### Patch Changes
+
+- Updated dependencies [085e855]
+  - @penumbra-zone/types@33.1.0
+  - @penumbra-zone/crypto-web@43.0.0
+  - @penumbra-zone/ui@14.0.1
+  - @penumbra-zone/wasm@47.0.0
+  - @penumbra-zone/perspective@56.0.0
+
+## 0.1.0
+
+### Minor Changes
+
+- 694319c: bump registry version
+
+### Patch Changes
+
+- Updated dependencies [4a51a46]
+- Updated dependencies [4a51a46]
+  - @penumbra-zone/ui@14.0.0
+
 ## 0.0.9
 
 ### Patch Changes