+++ packages/tailwind-config/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @penumbra-zone/tailwind-config
 
+## 3.1.39
+
+### Patch Changes
+
+- @penumbra-zone/ui-deprecated@19.0.1
+
+## 3.1.38
+
+### Patch Changes
+
+- Updated dependencies [4a51a46]
+  - @penumbra-zone/ui-deprecated@19.0.0
+
 ## 3.1.37
 
 ### Patch Changes